+++ scripts/penrose/tickevent/bedrock/bedrockvalidate.js
@@ -1,5 +1,5 @@
 import { world } from "mojang-minecraft";
-import { clearTickInterval, setTickInterval } from "../../../timer/scheduling.js";
+import { clearTickInterval, setTickInterval } from "../../../misc/scheduling.js";
 import config from "../../../data/config.js";
 import { crypto } from "../../../util.js";
