+++ scripts/penrose/tickevent/invalidsprint/invalidsprint_a.js
@@ -1,6 +1,6 @@
 import { world, MinecraftEffectTypes } from "mojang-minecraft";
 import { crypto, flag } from "../../../util.js";
-import { clearTickInterval, setTickInterval } from "../../../timer/scheduling.js";
+import { clearTickInterval, setTickInterval } from "../../../misc/scheduling.js";
 import config from "../../../data/config.js";
 
 const World = world;
