+++ scripts/penrose/tickevent/knockback/antikb_a.js
@@ -1,7 +1,7 @@
 import { world } from "mojang-minecraft";
 import { flag, crypto } from "../../../util.js";
 import config from "../../../data/config.js";
-import { clearTickInterval, setTickInterval } from "../../../timer/scheduling.js";
+import { clearTickInterval, setTickInterval } from "../../../misc/scheduling.js";
 
 const World = world;
